moving the updateCachedIcons to utils, and tiny fixes
