--- src/databases/Redis/index.ts
@@ -1,2 +1,0 @@
-export { default as redisClient } from './redis'
-export * as redisServices from './services.redis'
