+++ src/utils/index.ts
@@ -1,2 +1,4 @@
 export { default as FontFactory } from './FontFactory'
 export { default as SvgFactory } from './SvgFactory'
+export { default as updateDBIcons } from './updateDBIcons'
+export { default as updateCachedIcons } from './updateCachedIcons'
