+++ src/utils/updateDBIcons.ts
@@ -6,7 +6,8 @@ import { infoServices } from 'components/info'
 import { Logger, NodeLogger } from 'helpers'
 import { getEncodedLink, isNewIcon, getFilled, getOutlined, prepareUpdatedIcon } from './tools'
 import configs from 'configs'
-import { redisServices } from 'databases/Redis'
+import updateCachedIcons from './updateCachedIcons'
+
 // promise-style version of cmd.get
 const pcmdGet = util.promisify(cmd.get)
 
@@ -106,7 +107,7 @@ const updateDBIcons = async (notifiedByHook = false) => {
     updateIconsLogger.logInfo('', { message: 'The process of updating the icons is Finished' })
     console.timeEnd('Updating the icons Process')
     if (notifiedByHook) {
-      await redisServices.updateCachedIcons()
+      await updateCachedIcons()
     }
     // Add an info about the updated/added/deleted icons
     infoServices.createInfoDocument({ iconsAdded: namesOfNewIcons, iconsUpdated: namesOfUpdatedIcons, iconsDeleted: namesOfDeletedIcons })
